+++ src/api/endpoints.ts
@@ -1,19 +1,19 @@
+import CORE from '../../config/rfcx.json'
+
 export type RequestMethod = 'GET' | 'POST' | 'PUT' | 'PATCH' | 'DELETE'
 
 interface Endpoint {
   method: RequestMethod
   url: string
+  config?: any
 }
 
-const CORE = 'https://staging-api.rfcx.org'
-
-// Example: to be updated
 export const getStreams: Endpoint = {
   method: 'GET',
-  url: `${CORE}/streams`
+  url: `${CORE.rangerBaseApi}/streams`
 }
 
 export const getProjects: Endpoint = {
   method: 'GET',
-  url: `${CORE}/projects`
+  url: `${CORE.rangerBaseApi}/projects`
 }