+++ src/components/project-selector/project-selector.ts
@@ -1,8 +1,6 @@
 import { Options, Vue } from 'vue-class-component'
 import { Emit } from 'vue-property-decorator'
-
 import { OnClickOutside } from '@vueuse/components'
-
 import { ProjectModels } from '@/models'
 import { ROUTES_NAME } from '@/router'
 import { VuexService } from '@/services'
@@ -14,30 +12,32 @@ export default class ProjectSelectorComponent extends Vue {
   @VuexService.Project.projects.bind()
   projects!: ProjectModels.ProjectListItem[]
 
-  @VuexService.Project.selectedProject.bind()
-  selectedProject!: ProjectModels.ProjectListItem | undefined
-
-  public currentSelectedProject: ProjectModels.ProjectListItem | undefined = { ...VuexService.Project.selectedProject.get() }
+  public selectedProject: any = {}
 
   isSelectedProject (project: ProjectModels.ProjectListItem): boolean {
-    return project.id === this.currentSelectedProject?.id
+    return project.id === this.selectedProject?.id
+  }
+
+  getSelectedProject (): any {
+    this.selectedProject = this.projects.find(p=>p.id===this.$route.params.projectId)
   }
 
   setSelectedProject (project: ProjectModels.ProjectListItem): void {
-    console.log('\n\n----------setSelectedProject--------', project)
-    this.currentSelectedProject = project
+    this.selectedProject = project
   }
 
   async confirmedSelectedProject (): Promise<void> {
-    await VuexService.Project.selectedProject.set(this.currentSelectedProject)
-    this.closeProjectSelector()
-
-    const newProjectId = this.currentSelectedProject?.id
+    const newProjectId = this.selectedProject?.id
     if (newProjectId !== undefined) void this.$router.push({ name: ROUTES_NAME.guardians, params: { projectId: newProjectId } })
+    this.closeProjectSelector()
   }
 
   @Emit('closeProjectSelector')
   public closeProjectSelector (): boolean {
     return false
+  }
+
+  mounted (): void {
+    this.getSelectedProject()
   }
 }