+++ src/models/Map.ts
@@ -1,0 +1,6 @@
+import Mapbox from 'mapbox-gl'
+
+export interface MapboxOptions {
+  center: Mapbox.LngLatLike,
+  zoom: number
+}